+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+ * Added `outputDFS` to `&NAMLETKF` (#1105 and !1050)
  * Modified `LOCALSELECTIONOUTPUT` verbosity level (#1063 and !1014)
  * Added piecewise linear interpolation of the letkf horizontal localization radius (#1024 and !978)
  * Added `LOCALSELECTIONOUTPUT`, `MAXNUMLOCALOBSPERTYPE`, and `LOCALOBSSORTING` to `&NAMLETKF` (#998 and !921)